Authorization bug fix

+++ modules/admin/controllers/CategoryController.php
@@ -35,9 +35,14 @@ class CategoryController extends Controller
      */
     public function actionIndex()
     {
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
+
         $searchModel = new CategorySearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-
         return $this->render('index', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
@@ -51,6 +56,11 @@ class CategoryController extends Controller
      */
     public function actionView($id)
     {
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -63,6 +73,12 @@ class CategoryController extends Controller
      */
     public function actionCreate()
     {
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
+        
         $model = new Category();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -82,6 +98,13 @@ class CategoryController extends Controller
      */
     public function actionUpdate($id)
     {
+
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
+        
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -101,6 +124,12 @@ class CategoryController extends Controller
      */
     public function actionDelete($id)
     {
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
+        
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
@@ -115,6 +144,12 @@ class CategoryController extends Controller
      */
     protected function findModel($id)
     {
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
+        
         if (($model = Category::findOne($id)) !== null) {
             return $model;
         } else {
