+++ modules/admin/controllers/PostController.php
@@ -39,6 +39,11 @@ class PostController extends Controller
      */
     public function actionIndex()
     {
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
         $searchModel = new PostSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -55,6 +60,11 @@ class PostController extends Controller
      */
     public function actionView($id)
     {
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -67,6 +77,11 @@ class PostController extends Controller
      */
     public function actionCreate()
     {
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
         $model = new Post();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -82,6 +97,11 @@ class PostController extends Controller
 
     public function actionUpdate($id)
     {
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
 
         $model = $this->findModel($id);
 
@@ -104,6 +124,11 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
@@ -118,6 +143,11 @@ class PostController extends Controller
      */
     protected function findModel($id)
     {
+        $redirectUrl = '/web/';
+        if (Yii::$app->user->isGuest)
+        {
+            return Yii::$app->getResponse()->redirect($redirectUrl);
+        }
         if (($model = Post::findOne($id)) !== null) {
             return $model;
         } else {
